--- a/.github/workflows/tests.yml
+++ b/.github/workflows/tests.yml
@@ -21,9 +21,6 @@ jobs:
         with:
           go-version: ^1.16
 
-      - name: go get
-        run: cd /tmp && go get -v "github.com/${GITHUB_REPOSITORY}/${BRANCH}@${BRANCH}"
-
   lint:
     runs-on: ubuntu-latest
     if: ${{ !contains(github.ref, 'calendar') }}