--- a/hw10_program_optimization/go.mod
+++ b/hw10_program_optimization/go.mod
@@ -3,7 +3,7 @@ module github.com/raymanovg/otus_golang/hw10_program_optimization
 go 1.16
 
 require (
-	github.com/davecgh/go-spew v1.1.1 // indirect
+	github.com/json-iterator/go v1.1.12
 	github.com/stretchr/testify v1.7.0
 	gopkg.in/yaml.v3 v3.0.0-20210107192922-496545a6307b // indirect
 )