@raymanovg 2021-12-06
Benchmarks:
name             old time/op    new time/op    delta
GetDomainStat-8    2.72ms ± 0%    0.00ms ± 0%  -99.86%

name             old alloc/op   new alloc/op   delta
GetDomainStat-8    74.2kB ± 0%     4.5kB ± 0%  -93.88%

name             old allocs/op  new allocs/op  delta
GetDomainStat-8       738 ± 0%         7 ± 0%  -99.05%